Add trait requirements to `MeterMarker`
Add `Send` and `Sync` to `MeterMarker` and `MeterMarker::Field` to allow
them to be used safely across threads in preparation for creating the
effect component. Also add `NumAssignOps` and `Copy` trait requirement
on `MeterMarker::Field` to put bounds on `Meter` fields requiring them
to be compatible with numeric operations so that effects can change the
value of `Meter`s' fields.